Fix requestAnimationFrame spam/freeze-up from depth overflow
This commit fixes an error I was encountering on my local configuration
with long-living browser windows.

The problem is that a forever-running waitForElement call may result in
its requestAnimationFrame calls nesting past the 32-bit signed integer
maximum, at which point requestAnimationFrame will break for the document
(in this case, the document is the browser window itself, and cannot be
navigated away from) forever, doomed to return NS_ERROR_NOT_AVAILABLE.

I changed the default timeout behaviour from disabled to 10 seconds, which
is more than enough time for most elements to appear. Quite frankly, if any
code was written to wait for rare elements to be created using this method,
then they were poorly designed from the start and needed to be replaced
anyway. With that being said, I did not take any care into seeing what this
would affect. Echelon still works perfectly fine for me, without a single
noticeable error introduced by this change, so I am going to push it to
main.

Signed-off-by: Isabella Lulamoon <kawapure@gmail.com>